requirements.txt
Flask==2.0.1
torch==1.9.0
torchvision==0.10.0
numpy==1.21.0
Pillow==8.3.1
opencv-python==4.5.3.56
gunicorn==20.1.0
torch==1.9.0+cu111
torchvision==0.10.0+cu111
-f https://download.pytorch.org/whl/torch_stable.html